Example 1: Random patterns are created and used as stored patterns. Some noise is added to the patterns to perturb them. If one perturbed pattern and the original patterns are presented to the Hopfield, it correctly retrieves the stored random pattern.
Example 2: There are 10 different slopes. For each slope 100 samples are created by adding noise to the line of the respective slope. From the 100 samples one obtains the average which serves as the prototype.
If one perturbed pattern and the 10 protoypes are presented to the Hopfield, it CANNOT retrieves the stored random pattern. It will always select the prototype with the largest slope. This is natural behaviour
as the product inside the Hopfield softmax is a scalar product and since the prototype and the samples are not normalized, the largest value is always obtained for the prototyple with the largest slope.

Example 3: Instead of linear patterns, sine functions with 10 different frequencies are used. Again 100 samples are created for each frequency by adding noise. The average again serves as prototype. The major
difference to the above example is, that the scalar product with a random pattern will be equal to zero, due to the alternating signs of the sine function values. If one perturbed pattern and the 10 protoypes
are presented to the Hopfield, it retrieves the stored random pattern correctly.